the-wild-oasis : video 357

+++ the-wild-oasis/src/services/apiCabins.js
@@ -1,10 +1,84 @@
+// import supabase, { supabaseUrl } from './supabase';
+
+// export async function getCabins() {
+//   const { data, error } = await supabase.from('cabins').select('*');
+
+//   if (error) {
+//     console.log(error);
+//     throw new Error('Cabins could not be loaded');
+//   }
+
+//   return data;
+// }
+
+// export async function createEditCabin(newCabin, id) {
+//   //Check image when edit
+//   const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
+//   const imageName = `${Math.random()}-${
+//     newCabin.image.name
+//   }`.replaceAll('/', '');
+
+//   const imagePath = hasImagePath
+//     ? newCabin.image
+//     : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+//   //1. create/edit cabin
+//   let query = supabase.from('cabins');
+
+//   // A) Create
+//   if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+//   // B) Edit
+//   if (id)
+//     query = query
+//       .update({ ...newCabin, image: imagePath })
+//       .eq('id', id);
+
+//   const { data, error } = await query.select().single();
+
+//   if (error) {
+//     console.log(error);
+//     throw new Error('Cabins could not be deleted');
+//   }
+
+//   // 2.if success upload the image
+//   const { error: storageError } = await supabase.storage
+//     .from('cabin-images')
+//     .upload(imageName, newCabin.image);
+
+//   // 3. Delete the cabin if there was an error uploading image
+//   if (storageError) {
+//     await supabase.from('cabins').delete().eq('id', data.id);
+//     console.log(storageError);
+//     throw new Error(
+//       'Cabins image could not be uploaded and the cabin was not craeted'
+//     );
+//   }
+
+//   return data;
+// }
+
+// export async function deleteCabin(id) {
+//   const { data, error } = await supabase
+//     .from('cabins')
+//     .delete()
+//     .eq('id', id);
+
+//   if (error) {
+//     console.log(error);
+//     throw new Error('Cabins could not be deleted');
+//   }
+//   return data;
+// }
+
 import supabase, { supabaseUrl } from './supabase';
 
 export async function getCabins() {
   const { data, error } = await supabase.from('cabins').select('*');
 
   if (error) {
-    console.log(error);
+    console.error(error);
     throw new Error('Cabins could not be loaded');
   }
 
@@ -12,24 +86,22 @@ export async function getCabins() {
 }
 
 export async function createEditCabin(newCabin, id) {
-  //Check image when edit
   const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
   const imageName = `${Math.random()}-${
     newCabin.image.name
   }`.replaceAll('/', '');
-
   const imagePath = hasImagePath
     ? newCabin.image
     : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  //1. create/edit cabin
+  // 1. Create/edit cabin
   let query = supabase.from('cabins');
 
-  // A) Create
+  // A) CREATE
   if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
-  // B) Edit
+  // B) EDIT
   if (id)
     query = query
       .update({ ...newCabin, image: imagePath })
@@ -38,21 +110,23 @@ export async function createEditCabin(newCabin, id) {
   const { data, error } = await query.select().single();
 
   if (error) {
-    console.log(error);
-    throw new Error('Cabins could not be deleted');
+    console.error(error);
+    throw new Error('Cabin could not be created');
   }
 
-  // 2.if success upload the image
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from('cabin-images')
     .upload(imageName, newCabin.image);
 
-  // 3. Delete the cabin if there was an error uploading image
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from('cabins').delete().eq('id', data.id);
-    console.log(storageError);
+    console.error(storageError);
     throw new Error(
-      'Cabins image could not be uploaded and the cabin was not craeted'
+      'Cabin image could not be uploaded and the cabin was not created'
     );
   }
 
@@ -66,8 +140,9 @@ export async function deleteCabin(id) {
     .eq('id', id);
 
   if (error) {
-    console.log(error);
-    throw new Error('Cabins could not be deleted');
+    console.error(error);
+    throw new Error('Cabin could not be deleted');
   }
+
   return data;
 }
